+++ src/BasicEnvironment.cpp
@@ -94,14 +94,27 @@ void BasicEnvironment::insert_agent(BasicAgent *agent){
     }
 }
 
+
+void BasicEnvironment::insert_predator(Predator *predator){
+    predator_ptrs.push_back(predator);
+    environment_tiles[predator->y_pos - 1][predator->x_pos - 1] = 'P';
+
+    // Updates each agent's knowledge of environment when a new agent is inserted into one
+    for (Predator *pred: predator_ptrs) {
+        pred->environment_tiles = environment_tiles;
+        pred->is_environ_set = true;
+    }
+}
+
+
 void BasicEnvironment::update() {
 
     time_instance += 1;
 
-    // Resets any previous agent icon on tiles from 'A' to '-'
+    // Resets any previous agent icon on tiles from 'A' or 'P' to '-'
     for (int i=0; i<environment_tiles.size(); i++) {
         for (int j=0; j<environment_tiles[i].size(); j++) {
-            if (environment_tiles[i][j] == 'A') {
+            if (environment_tiles[i][j] == 'A' or environment_tiles[i][j] == 'P') {
                 environment_tiles[i][j] = '-';
             }
         }
@@ -128,6 +141,19 @@ void BasicEnvironment::update() {
             }
         }
         agnt->energies_consumed.clear();
+    }
+    // Adds the new predators onto the tiles
+    for (Predator *pred: predator_ptrs) {
+        if (environment_tiles[pred->y_pos-1][pred->x_pos-1] == 'A') {
+            cout << "Predator '" << pred->name << "' found agent at (" << pred->x_pos << ", " << pred->y_pos << ") and has eaten it!" << endl;
+            // Finds the agent which has been eaten and set it to be dead
+            for (BasicAgent *agnt: agent_ptrs) {
+                if (agnt->x_pos == pred->x_pos and agnt->y_pos == pred->y_pos) {
+                    agnt->is_dead = true;
+                }
+            }
+        }
+        environment_tiles[pred->y_pos-1][pred->x_pos-1] = 'P';
     }
 }